add equipment page under manage page
